Add a pre-commit hook to the mix
This runs the tests and doesn't let you commit till the tests pass in
both python2 and python3. To enable it run::

  $ ln -sv ../../.pre-commit.hook .git/hooks/pre-commit

from the root of this git repo.

Will generally be quiet unless a test fails and we have to pop the stash
back in, then we say what was changed and what failed.